[BEAM-12629] Make source tests spark version agnostic and move them back to common spark module
